+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.4]
 ### Added
 - Support cases where `type` is an array in JSONoid object conversion
 - Support enums in JSONoid object conversion
@@ -140,7 +142,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-05-31
 - Initial release
 
-[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.3...HEAD
+[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.4...HEAD
+[0.5.4]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.0...v0.5.1